Pull out AccessorDeclarationNodeConverter
There's probably some general block statement logic that could be pulled out too
For now I'm just trying to grab the biggest chunks to have the biggest impact on reducing line count
